- Implemented share feature

+++ app/src/main/java/de/stubbe/jaem_client/utils/EncryptionExtension.kt
@@ -51,7 +51,7 @@ fun Short.toByteArray(): ByteArray {
 fun Long.toByteArray(): ByteArray {
     return ByteBuffer.allocate(ULong.SIZE_BYTES)
         .order(ByteOrder.BIG_ENDIAN)
-        .putLong(this.toLong())
+        .putLong(this)
         .array()
 }
